drivers/leds/leds-s2mu106: Custom sysfs node for flashlight
This creates a custom sysfs node located at sys/devices/virtual/camera/flash/torch_brightness_lvl
for the supported LED. This sysfs node can be used to control flashlight brightness on
AOSP ROMs or recoveries.

By default, the brightness of flashlight is set to maximum, that is, it will fix low
flashlight issue on AOSP ROMs, GSIs and recoveries.

This sysfs node can also override the stock behaviour of ONEUI's flashlight controller
(only if uncommented in the driver code).

The flashlight brightness range is from 0 to 10.

Change-Id: I8d7a2d0591a8ab5ab5374cf7f58b8d94e73d791b
Signed-off-by: Chatur27 <jasonbright2709@gmail.com>